+++ samples/a2a/langchain-weather-agent/.dockerignore
@@ -1,0 +1,14 @@
+venv/
+.venv/
+__pycache__/
+*.pyc
+*.pyo
+*.pyd
+.pytest_cache/
+.mypy_cache/
+.env
+.env.local
+*.egg-info/
+dist/
+build/
+
