+++ samples/a2a/langchain-weather-agent/env.example
@@ -1,0 +1,17 @@
+# Azure OpenAI Configuration
+# Copy this file to .env and fill in your values
+
+AZURE_OPENAI_API_KEY=your-azure-openai-api-key-here
+AZURE_API_BASE=https://your-resource.openai.azure.com
+LLM_MODEL_NAME=gpt-4o-mini
+AZURE_API_VERSION=2024-04-01-preview
+
+# Common deployment names:
+# - gpt-4o-mini (recommended, fast and cheap)
+# - gpt-4o (newest GPT-4)
+# - gpt-35-turbo (GPT-3.5)
+# - gpt-4 or gpt-4-turbo (older GPT-4)
+#
+# Note: LLM_MODEL_NAME should match your Azure OpenAI deployment name,
+# not necessarily the base model name. Check your Azure portal for available deployments.
+
