+++ samples/a2a/langchain-weather-agent/src/langchain_weather_agent/__init__.py
@@ -1,0 +1,4 @@
+"""LangChain weather agent A2A server."""
+
+__version__ = "1.0.0"
+
