+++ samples/a2a/langchain-weather-agent/src/langchain_weather_agent/__main__.py
@@ -1,0 +1,7 @@
+"""Main entry point for LangChain weather agent."""
+
+from langchain_weather_agent.a2a_server import main
+
+if __name__ == "__main__":
+    main()
+
