--- samples/agent-hosting/hosted-langchain-agents/.env.sample
@@ -1,7 +1,0 @@
-# Azure OpenAI configuration for LangChain agents
-AZURE_OPENAI_API_KEY=your_azure_openai_key_here
-AZURE_API_BASE=https://lxo.openai.azure.com
-AZURE_API_VERSION=2024-12-01-preview
-
-# Model deployment name
-LLM_MODEL_NAME=gpt-4o
