--- samples/agent-hosting/hosted-langchain-agents/chart/Chart.yaml
@@ -1,9 +1,0 @@
-apiVersion: v2
-name: hosted-langchain-agents
-description: Hosted LangChain agents sample with A2A server integration
-type: application
-version: 0.1.0
-appVersion: 0.1.39
-annotations:
-  ark.mckinsey.com/service: hosted-langchain-agents
-  ark.mckinsey.com/resources: service,a2aserver
