--- samples/agent-hosting/hosted-langchain-agents/chart/templates/a2aserver.yaml
@@ -1,17 +1,0 @@
-{{- if .Values.a2aserver.enabled }}
-apiVersion: ark.mckinsey.com/v1prealpha1
-kind: A2AServer
-metadata:
-  name: {{ .Values.a2aserver.name }}
-  {{- with .Values.global.annotations }}
-  annotations:
-    {{- toYaml . | nindent 4 }}
-  {{- end }}
-spec:
-  address:
-    valueFrom:
-      serviceRef:
-        name: {{ .Values.service.name }}
-        port: http
-  description: "LangChain weather agent hosted via A2A protocol"
-{{- end }}
