--- samples/agent-hosting/hosted-langchain-agents/chart/templates/deployment.yaml
@@ -1,75 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: {{ .Values.app.name }}
-  labels:
-    app: {{ .Values.app.name }}
-  {{- with .Values.global.annotations }}
-  annotations:
-    {{- toYaml . | nindent 4 }}
-  {{- end }}
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: {{ .Values.app.name }}
-  template:
-    metadata:
-      labels:
-        app: {{ .Values.app.name }}
-      {{- with .Values.global.annotations }}
-      annotations:
-        {{- toYaml . | nindent 8 }}
-      {{- end }}
-    spec:
-      serviceAccountName: {{ .Values.serviceAccount.name }}
-      automountServiceAccountToken: false
-      containers:
-      - name: {{ .Values.app.name }}
-        image: "{{ .Values.app.image.repository }}:{{ .Values.app.image.tag | default .Chart.AppVersion }}"
-        imagePullPolicy: {{ .Values.app.image.pullPolicy }}
-        ports:
-        - containerPort: {{ .Values.service.targetPort }}
-        {{- if .Values.app.env }}
-        env:
-          {{- range .Values.app.env }}
-          - name: {{ .name }}
-            {{- if .value }}
-            value: {{ .value | quote }}
-            {{- else if .valueFrom }}
-            valueFrom:
-              {{- toYaml .valueFrom | nindent 14 }}
-            {{- end }}
-          {{- end }}
-        {{- end }}
-        {{- if .Values.app.resources }}
-        resources:
-          {{- toYaml .Values.app.resources | nindent 10 }}
-        {{- end }}
-        # Using A2A agent card endpoint for probes (interim solution)
-        # NOTE: A dedicated probe type would be better for A2A protocol compliance
-        startupProbe:
-          httpGet:
-            path: /health
-            port: {{ .Values.service.targetPort }}
-          initialDelaySeconds: 1
-          periodSeconds: 10
-          timeoutSeconds: 5
-          failureThreshold: 30
-        readinessProbe:
-          httpGet:
-            path: /.well-known/agent.json
-            port: {{ .Values.service.targetPort }}
-          initialDelaySeconds: 2
-          periodSeconds: 10
-          timeoutSeconds: 5
-          successThreshold: 1
-          failureThreshold: 3
-        livenessProbe:
-          httpGet:
-            path: /health
-            port: {{ .Values.service.targetPort }}
-          initialDelaySeconds: 0
-          periodSeconds: 30
-          timeoutSeconds: 5
-          failureThreshold: 3
