--- samples/agent-hosting/hosted-langchain-agents/chart/templates/serviceaccount.yaml
@@ -1,10 +1,0 @@
-{{- if .Values.serviceAccount.create }}
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: {{ .Values.serviceAccount.name }}
-  {{- with .Values.global.annotations }}
-  annotations:
-    {{- toYaml . | nindent 4 }}
-  {{- end }}
-{{- end }}
