--- samples/agent-hosting/hosted-langchain-agents/chart/values.yaml
@@ -1,57 +1,0 @@
-# Hosted LangChain Agents chart values
-
-# Global annotations added to all resources
-global:
-  annotations:
-    ark.mckinsey.com/service: hosted-langchain-agents
-
-# Application configuration
-app:
-  name: hosted-langchain-agents
-  image:
-    repository: hosted-langchain-agents
-    tag: latest
-    pullPolicy: IfNotPresent
-  
-  # Resource configuration
-  resources:
-    requests:
-      memory: "256Mi"
-      cpu: "250m"
-    limits:
-      memory: "512Mi"
-      cpu: "500m"
-  
-  # Environment variables
-  env:
-    - name: LLM_MODEL_NAME
-      value: "gpt-4o"
-    - name: AZURE_API_VERSION
-      value: "2024-12-01-preview"
-    - name: AZURE_OPENAI_API_KEY
-      valueFrom:
-        secretKeyRef:
-          name: langchain-azure-openai-secret
-          key: token
-    - name: AZURE_API_BASE
-      valueFrom:
-        secretKeyRef:
-          name: langchain-azure-openai-secret
-          key: base-url
-
-# Service configuration
-service:
-  name: hosted-langchain-agents
-  type: ClusterIP
-  port: 80
-  targetPort: 8000
-
-# A2A Server configuration
-a2aserver:
-  name: hosted-langchain-agents
-  enabled: true
-
-# Service account configuration
-serviceAccount:
-  name: hosted-langchain-agents-sa
-  create: true
