--- samples/agent-hosting/hosted-langchain-agents/setup.cfg
@@ -1,3 +1,0 @@
-[flake8]
-max-line-length = 120
-extend-ignore = E203, W503
