--- samples/agent-hosting/hosted-langchain-agents/src/hosted_langchain_agents/__init__.py
@@ -1,1 +1,0 @@
-"""LangChain agents hosted via A2A protocol."""
