--- samples/agent-hosting/hosted-langchain-agents/src/hosted_langchain_agents/__main__.py
@@ -1,6 +1,0 @@
-"""Main entry point for hosted LangChain agents."""
-
-from hosted_langchain_agents.a2a_server import main
-
-if __name__ == "__main__":
-    main()
